Ensure available pegs are accurately displayed
If a user changes assigns the "red" peg to position 0 in their guess,
red should be removed from the available pegs. If the user then assigns
"yellow" to position 0, red should reappear as available (and yellow
should be removed).

Available pegs are always the difference between the list of all
available colors and the user's current guess.