Show guess feedback in the form of KeyPegs
Black KeyPeg: the user's guess contains 1 peg which is the same
color/position as the secret code.

White KeyPeg: the user's guess contains 1 peg which is the same color as
the secret code, but in the wrong position.

Adds new KeyPeg and KeyPegsList components. The logic for whether
evaluating the guess currently lives in the KeyPegsList component - not
sure it makes sense at this point to actually store that in app state...